cockroach-db/requirements.txt
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
sqlalchemy-cockroachdb==2.0.2
python-dotenv==1.0.1
flask==3.0.3
locust==2.29.1
